refactor sync mod
